TASK: Move `TranslationRepository` into `model` module
