+++ packages/neos-ui-i18n/src/model/TranslationRepository.spec.ts
@@ -7,7 +7,9 @@
  * information, please view the LICENSE file which was distributed with this
  * source code.
  */
-import {Locale, Translation, TranslationAddress} from '../model';
+import {Locale} from './Locale';
+import {Translation} from './Translation';
+import {TranslationAddress} from './TranslationAddress';
 
 import {TranslationRepository} from './TranslationRepository';
