+++ packages/neos-ui-i18n/src/model/TranslationRepository.ts
@@ -7,12 +7,9 @@
  * information, please view the LICENSE file which was distributed with this
  * source code.
  */
-import {
-    Locale,
-    Translation,
-    TranslationDTO,
-    type TranslationAddress
-} from '../model';
+import {Locale} from './Locale';
+import {Translation, type TranslationDTO} from './Translation';
+import type {TranslationAddress} from './TranslationAddress';
 
 export type TranslationsDTO = Record<string, Record<string, Record<string, TranslationDTO>>>;
